Changed Mats from double to float type.

+++ utils.h
@@ -5,6 +5,7 @@
 #include <opencv2/highgui.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
 #include <iostream>
+#include <iomanip>
 
 using namespace cv;
 
@@ -14,6 +15,6 @@ void show(Mat &mat, String name);
 void print_val(Mat &mat, String name);
 void integralImage(Mat &i, Mat &ii, Mat &S, int x, int y);
 Mat mat_box_filter(Mat &srcMat, int r);
-double S(Mat &i, Mat &sMat, int x, int y, int offset);
-double ii(Mat &i, Mat &iiMat, Mat &sMat, int x, int y, int offset);
+float S(Mat &i, Mat &sMat, int x, int y, int offset);
+float ii(Mat &i, Mat &iiMat, Mat &sMat, int x, int y, int offset);
 #endif //GUIDED_FILTER_CUDA_UTILS_H
